+++ R/installGoesGrids.R
@@ -1,8 +1,6 @@
 #' @export
 #' 
 #' @title Create GOES grids in satelliteDataDir
-#' @param showProgress logical specify whether to show progress bar while 
-#' building grids
 #' @description Creates data files with geolocation information for GOES-16 
 #' (East) and GOES-17 (West) satellite products.  Checks for the presence of
 #' GOES 16 & GOES 17 .nc (NetCDF) files in /inst/extdata. If present, it
@@ -10,14 +8,11 @@
 #' GOES East and West grids in the directory previously set with 
 #' \code{setSatelliteDataDir()}.
 #' 
-#' JON: What do you want to do if the .nc files DON'T exist?
 #' TODO: Function currently generates warinings:
 #' "In sqrt(b^2 - 4 * a * c) : NaNs produced"
 #' Check that we can safely ignore them, or even suppressWarningMessages({ ... })
 #' 
-#' 
-installGoesGrids <- function(
-) {
+installGoesGrids <- function() {
   
   outputDir <- getSatelliteDataDir()
   
@@ -26,9 +21,13 @@ installGoesGrids <- function(
   filename <- "goesEastGrid.rda"
   G16_filepath <- file.path(outputDir, filename)
   
-  if ( !file.exists(G16_filepath) ) {
-    cat(sprintf("Creating %s", G16_filepath), "\n")
+  if ( file.exists(G16_filepath) ) {
     
+    message(sprintf("Found %s", G16_filepath))
+    
+  } else {
+    
+    message(sprintf("Creating %s ...", G16_filepath))
     
     # Get a NetCDF handle for the package internal GOES-16 dataset
     nc_filepath <- system.file(
@@ -36,7 +35,11 @@ installGoesGrids <- function(
       "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc", 
       package = "MazamaSatelliteUtils"
     )
-    nc <- ncdf4::nc_open(nc_filepath)
+    if ( file.exists(nc_filepath) ) {
+      nc <- ncdf4::nc_open(nc_filepath)
+    } else {
+      stop(paste0("GOES-16 example file missing from package data."))
+    }
     
     # Get the projection information
     projection <- goesaodc_getProjection(nc)
@@ -46,7 +49,6 @@ installGoesGrids <- function(
     ncol <- length(ncvar_get(nc, varid = "y"))
     
     # Create a tibble where each AOD value has an associated location
-    cat(sprintf("%s", "................"), "\n")
     coordGrid <- goesaodc_getCoordGrid(nc)
     longitude <- coordGrid$lon
     latitude <- coordGrid$lat
@@ -59,7 +61,7 @@ installGoesGrids <- function(
     )  
     
     save(goesEastGrid, file = G16_filepath)
-    cat(sprintf("%s", "Done"), "\n")
+    message(paste0("... done!"))
     
   }
   
@@ -68,8 +70,13 @@ installGoesGrids <- function(
   filename <- "goesWestGrid.rda"
   G17_filepath <- file.path(outputDir, filename)
   
-  if ( !file.exists(G17_filepath) ) {
-    cat(sprintf("Creating %s", G17_filepath), "\n")
+  if ( file.exists(G17_filepath) ) {
+    
+    message(sprintf("Found %s", G17_filepath))
+    
+  } else {
+    
+    message(sprintf("Creating %s ...", G17_filepath))
     
     # Get a NetCDF handle for the package internal GOES-17 dataset
     nc_filepath <- system.file(
@@ -77,8 +84,12 @@ installGoesGrids <- function(
       "OR_ABI-L2-AODC-M6_G17_s20192491826196_e20192491828569_c20192491830494.nc", 
       package = "MazamaSatelliteUtils"
     )
-    nc <- ncdf4::nc_open(nc_filepath)
-    
+    if ( file.exists(nc_filepath) ) {
+      nc <- ncdf4::nc_open(nc_filepath)
+    } else {
+      stop(paste0("GOES-17 example file missing from package data."))
+    }
+
     # Get the projection information
     projection <- goesaodc_getProjection(nc)
     
@@ -87,7 +98,6 @@ installGoesGrids <- function(
     ncol <- length(ncvar_get(nc, varid = "y"))
     
     # Create a tibble where each AOD value has an associated location
-    cat(sprintf("%s", "................"), "\n")
     coordGrid <- goesaodc_getCoordGrid(nc)
     longitude <- coordGrid$lon
     latitude <- coordGrid$lat
@@ -100,7 +110,7 @@ installGoesGrids <- function(
     ) 
     
     save(goesWestGrid, file = G17_filepath)
-    cat(sprintf("%s", "Done"), "\n")
+    message(paste0("... done!"))
     
   }